Refactor code so guests are created and filtered using a unique id
Course creates and filters guests using the map and filter index.
This causes problems when removing a guest that is already filtered
according to rsvp response. A better solution is for each guest to
have a unique id, which solves any problems that previously occured
when removing and adding guests